pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "email-validator>=2.2.0",
    "flask>=3.0.3",
    "flask-sqlalchemy>=3.1.1",
    "psycopg2-binary>=2.9.9",
    "vllm>=0.6.2",
    "gitpython>=3.1.43",
    "grpcio>=1.66.2",
    "grpcio-tools>=1.66.2",
    "transformers>=4.45.2",
    "torch>=2.4.0",
]
